complete migration for track table
