+++ packages/raga-app/src/client/components/chrome/appChrome.module.scss
@@ -17,7 +17,7 @@
   flex-direction: row;
   gap: 5px;
   align-items: center;
-  height: 24px;
+  height: 30px;
   text-overflow: ellipsis;
   white-space: nowrap;
